refactor(tripBooking): folder structure refactor
