+++ src/pages/TripBookingPage/PassengerDetailsForm/PassengerDetailsForm.tsx
@@ -1,9 +1,10 @@
 import { Formik, Form, FieldArray, FastField } from 'formik';
-import { Box, Button, Grid, MenuItem, Paper, Typography } from '@mui/material';
+import { Box, Grid, MenuItem, Paper, Typography } from '@mui/material';
 import getValidationSchema from './validation';
 import { Select, TextField } from 'formik-mui';
 import { useTranslation } from 'react-i18next';
-import { Checkout } from '../../../../components/ui/TripCard/AccordionDetails/components/Checkout';
+import { Checkout } from '../../../components/ui/TripCard/AccordionDetails/components/Checkout';
+import { StyledButton } from '../../../components/Button/Button.styled';
 
 const PassengerDetailsForm = () => {
     const { t } = useTranslation('passengerDetails');
@@ -68,6 +69,7 @@ const PassengerDetailsForm = () => {
                                                 name={`passengers.${index}.fullName`}
                                                 id={`passengers.${index}.fullName`}
                                                 label={t('fullName')}
+                                                required
                                                 helperText=" "
                                             />
                                         </Grid>
@@ -79,6 +81,7 @@ const PassengerDetailsForm = () => {
                                                 id={'age'}
                                                 name={`passengers.${index}.age`}
                                                 label={t('age')}
+                                                required
                                                 helperText=" "
                                             />
                                         </Grid>
@@ -94,7 +97,10 @@ const PassengerDetailsForm = () => {
                                                 id={`passengers.${index}.gender`}
                                                 label={t('gender')}
                                                 labelId={`passengers.${index}.gender-simple`}
-                                                helperText=" "
+                                                required
+                                                formHelperText={{
+                                                    children: ' ',
+                                                }}
                                             >
                                                 {t('genderValues', {
                                                     returnObjects: true,
@@ -114,14 +120,14 @@ const PassengerDetailsForm = () => {
                         }
                     </FieldArray>
 
-                    <Grid container sx={{ alignItems: 'center' }} mb={6}>
+                    <Grid container alignItems="center" mb={6} spacing={1.5}>
                         <Grid item xs={12} sm={9}>
                             <Checkout noOfSeats={3} farePerSeat={1200} />
                         </Grid>
                         <Grid item xs={12} sm={3} ml="auto">
-                            <Button variant="contained" type="submit" fullWidth>
+                            <StyledButton type="submit" fullWidth>
                                 {t('checkout')}
-                            </Button>
+                            </StyledButton>
                         </Grid>
                     </Grid>
                 </Form>
